refactor express server running in the same process as in main
