+++ src/expressServer/routes.ts
@@ -1,5 +1,5 @@
-const express = require('express');
-const fs = require('fs');
+import express from 'express';
+import fs from 'fs';
 
 const router = express.Router();
 
@@ -7,7 +7,7 @@ router.get('/', (_, res) => {
   res.send('Server Running');
 });
 
-function streamVideo(req, res) {
+const streamVideo: (req: any, res: any) => void = (req, res) => {
   const encodeFilePath = req.params.name;
   const sourcePath = Buffer.from(encodeFilePath, 'base64').toString('utf-8');
   const stat = fs.statSync(sourcePath);
@@ -45,8 +45,8 @@ function streamVideo(req, res) {
     res.writeHead(200, head);
     fs.createReadStream(sourcePath).pipe(res);
   }
-}
+};
 
 router.get('/video/:name', streamVideo);
 
-module.exports = router;
+export default router;
