+++ src/expressServer/server.ts
@@ -1,0 +1,7 @@
+import express from 'express';
+import router from './routes';
+
+const app = express();
+app.use('/', router);
+
+export default app;
