+++ src/main/main.ts
@@ -32,7 +32,6 @@ export default class AppUpdater {
 
 let mainWindow: BrowserWindow | null = null;
 let pyServer: ChildProcess | null = null;
-let expressServer: ChildProcess | null = null;
 
 dotenv.config();
 
@@ -121,19 +120,7 @@ const createWindow = async () => {
       });
     }
 
-    expressServer = startExpressServer();
-
-    if (expressServer !== null) {
-      get(`http://localhost:${process.env.EXPRESS_PORT}`, (res) => {
-        if (res.statusCode === 200) {
-          console.log('Express server is running');
-        } else {
-          throw new Error(
-            `Express server has an error, response to a get '/' expected code 200 got ${res.statusCode}`
-          );
-        }
-      });
-    }
+    startExpressServer();
   });
 
   mainWindow.on('closed', () => {
